% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_plotify.R
\name{gantt_plotify}
\alias{gantt_plotify}
\title{Create a Gantt chart based on a valid `ganttrify` project}
\usage{
gantt_plotify(project = NULL, projectified = NULL, ...)
}
\arguments{
\item{project}{Defaults to NULL. Expected to be a valid `ganttrify` project data frame. Ignored if `projectified` given.}

\item{projectified}{Defaults to NULL. Expected to have the exact same format as the data frames generated with [gantt_projectify()]. Takes precedence of `project`.}

\item{...}{Arguments passed to [gantt_projectify()]. Ignored if `projectified` is not NULL}
}
\value{
A ggplot2 object, a nice-looking Gantt chart.
}
\description{
Pre-processed or customised inputs welcome.
}
\examples{
gantt_plotify(project = ganttrify::test_project)
}
