% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttrify.R
\name{gantt_points}
\alias{gantt_points}
\title{Add points to the gantt chart}
\usage{
gantt_points(x, shape = 21, size = 1, colour = "black", fill = "black")
}
\arguments{
\item{x}{The points data frame (with an 'activity' and a 'date' column)}

\item{shape}{The shape of the points}

\item{size}{The size of the points}

\item{colour}{The colour of the points}

\item{fill}{The fill of the points}
}
\description{
Add points to the gantt chart
}
\examples{
p <- data.frame(activity = "1.1. That admin activity", 
                date = as.Date("2020-12-12"))
ganttrify(ganttrify::test_project) + gantt_points(p)

}
