% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_projectify.R
\name{gantt_projectify}
\alias{gantt_projectify}
\title{Process a `ganttrify` project and make it amenable to processing with ggplot2}
\usage{
gantt_projectify(
  project,
  by_date = FALSE,
  exact_date = FALSE,
  project_start_date = Sys.Date(),
  colour_palette = wesanderson::wes_palette("Darjeeling1"),
  size_wp = 6,
  hide_wp = FALSE,
  wp_label_bold = TRUE,
  size_activity = 4,
  hide_activity = FALSE,
  label_wrap = FALSE,
  alpha_wp = 1,
  alpha_activity = 1,
  line_end = NULL,
  line_end_wp = "round",
  line_end_activity = "butt",
  month_breaks = 1
)
}
\arguments{
\item{project}{A data frame. See `ganttrify::test_project` for an example.}

\item{by_date}{Logical, defaults to FALSE If FALSE, the the start and end
columns in the data frame should correspond to month numbers from the
beginning of the project. If TRUE, dates in the format ("2020-10" or
"2020-10-01") should be given.}

\item{exact_date}{Logical, defaults to FALSE. If FALSE, then periods are
always understood to include full months. If FALSE, then exact dates can be
given.}

\item{project_start_date}{The date when the project starts. It can be a date,
or a string in the format "2020-03" or "2020-03-01". Ignored if
`month_number_date` is set to FALSE.}

\item{colour_palette}{A character vector of colours or a colour palette. If
necessary, colours are recycled as needed. Defaults to
`wesanderson::wes_palette("Darjeeling1")`. For more palettes, consider also
the `MetBrewer` package, e.g. `colour_palette =
MetBrewer::met.brewer("Lakota")`. Colours can be passed as a vector of hex
codes (e.g. `colour_palette = c("#6ACCEA", "#00FFB8", "#B90000",
"#6C919C")`)}

\item{size_wp}{Numeric, defaults to 6. It defines the thickness of the line
used to represent WPs.}

\item{hide_wp}{Logical, defaults to FALSE. If TRUE, the lines of the WP are
hidden and only activities are shown.}

\item{wp_label_bold}{Logical, defaults to \code{TRUE}. If \code{TRUE}, the
label for working packages is set to bold face, while activities remain
plain. Set to \code{FALSE} to keep have all labels in plain face or for
further customisation (`ganttrify` offers basic markdown support).}

\item{size_activity}{Numeric, defaults to 4. It defines the thickness of the
line used to represent activities.}

\item{hide_activity}{Logical, defaults to FALSE. If TRUE, the lines of
activities are hidden and only activities are shown.}

\item{label_wrap}{Defaults to FALSE. If given, must be numeric, referring to
the number of characters per line allowed in the labels of projects and
activities, or logical (if set to TRUE, it will default to 32). To be used
when labels would otherwise be excessively long.}

\item{alpha_wp}{Numeric, defaults to 1. Controls transparency of the line
used to represent WPs.}

\item{alpha_activity}{Numeric, defaults to 1. Controls transparency of the
line used to represent activities.}

\item{line_end}{Character, defaults to NULL. If given, takes precedence over
`line_end_wp` and `line_end_activity` and applies the value to both. One of
"round", "butt", "square". Controls line ends.}

\item{line_end_wp}{Character, defaults to "round". One of "round", "butt",
"square". Controls line ends.}

\item{line_end_activity}{Character, defaults to "butt". One of "round",
"butt", "square". Controls line ends.}

\item{month_breaks}{Numeric, defaults to 1. It defines if labels for all
months are shown or only once every x months. Useful for longer projects.}
}
\value{
A data frame, expected input for [gantt_plotify()].
}
\description{
Process a `ganttrify` project and make it amenable to processing with ggplot2
}
\examples{
gantt_projectify(project = ganttrify::test_project)
}
