% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_verify.R
\name{gantt_verify}
\alias{gantt_verify}
\title{Check the consistency of the input project data frame}
\usage{
gantt_verify(project, by_date = FALSE, exact_date = FALSE)
}
\arguments{
\item{project}{A data frame. See `ganttrify::test_project` for an example.}

\item{by_date}{Logical, defaults to FALSE If FALSE, the the start and end
columns in the data frame should correspond to month numbers from the
beginning of the project. If TRUE, dates in the format ("2020-10" or
"2020-10-01") should be given.}

\item{exact_date}{Logical, defaults to FALSE. If FALSE, then periods are
always understood to include full months. If FALSE, then exact dates can be
given.}
}
\value{
A data frame (a tibble) that is consistent with the format expected by [ganttrify()].
}
\description{
Check the consistency of the input project data frame, return meaningful errors or warnings if something is not quite right
}
\examples{
gantt_verify(project = ganttrify::test_project)
}
