% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttrify.R
\name{ganttrify}
\alias{ganttrify}
\title{Create Gantt charts from a data frame}
\usage{
ganttrify(
  project,
  spots = NULL,
  by_date = FALSE,
  exact_date = FALSE,
  project_start_date = Sys.Date(),
  colour_palette = wesanderson::wes_palette("Darjeeling1"),
  font_family = "Roboto Condensed",
  mark_quarters = FALSE,
  size_wp = 6,
  size_activity = 4,
  size_text_relative = 1,
  month_number = TRUE,
  colour_stripe = "lightgray"
)
}
\arguments{
\item{project}{A data frame. See `ganttrify::test_project` for an example.}

\item{spots}{A data frame. See `ganttrify::test_spots` for an example.}

\item{by_date}{Logical, defaults to FALSE If FALSE, the the start and end columns in the data frame should correspond to month numbers from the beginning of the project. If TRUE, dates in the format ("2020-10" or "2020-10-01") should be given.}

\item{exact_date}{Logical, defaults to FALSE. If FALSE, then periods are always understood to include full months. If FALSE, then exact dates can be given.}

\item{project_start_date}{The date when the project starts. It can be a date, or a string in the format "2020-03" or "2020-03-01". Ignored if `month_number` is set to FALSE.}

\item{colour_palette}{A character vector of colours or a colour palette.}

\item{size_text_relative}{Numeric, defaults to 1. Changes the size of all textual elements relative to their default size. If you set this to e.g. 1.5 all text elements will be 50\% bigger.}

\item{month_number}{Logical, defaults to TRUE. If TRUE, it included month numbering on top of the chart.}

\item{colour_stripe}{Character, defaults to "lightgray". This is the stripe colour in the background used in alternate months.}
}
\value{
A processed data frame ready to be turned into a Gantt chart.
}
\description{
Creates Gantt charts with ggplot2.
}
\examples{
ganttrify(ganttrify::test_project)

}
