% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ganttrify.R
\name{shiny_ganttrify}
\alias{shiny_ganttrify}
\title{Run a web app to customise Gantt charts}
\usage{
shiny_ganttrify()
}
\value{
Nothing, used for its side effects.
}
\description{
Creates Gantt with `ganttrify` without coding.
}
\examples{
ganttrify(ganttrify::test_project)

}
