% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttrify.R
\name{ganttrify}
\alias{ganttrify}
\title{Illustration of crayon colors}
\usage{
ganttrify(df, spots = NULL, start_date = Sys.Date(),
  colour_palette = wesanderson::wes_palette("Darjeeling1"),
  font_family = "Roboto Condensed", mark_quarters = FALSE,
  size_wp = 6, size_activity = 4, size_text_relative = 1)
}
\arguments{
\item{df}{A data frame.}

\item{start_date}{The date when the project starts. It can be a date, or a string in the format "2020-03" or "2020-03-01".}

\item{colour_palette}{A character vector of colours or a colour palette.}

\item{size_text_relative}{Numeric, defaults to 1. Changes the size of all textual elements relative to their default size. If you set this to e.g. 1.5 all text elements will be 50% bigger.}
}
\value{
A processed data frame ready to be turned into a Gantt chart.
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
ganttrify()

}
